Bug 1491098 - Add gmp moz.build files (WIP)
Most of the moz.build files are created manually by analyzing what files
are built during a normal './configure && make' build of the upstream
libgmp library. The exceptions are mpn/moz.build and mpn2/moz.build,
which are auto-generated from config_links.py and the config.status file
generated by libgmp's configure.

The mpn directory is special because:

1) It contains platform-dependent code
2) It compiles some .c files multiple times with different -D flags
  (eg: sec_aors_1.c is compiled as sec_sub_1 and sec_add_1)
3) It contains .asm files, which are compiled by first preprocessing the
.asm with an m4 wrapper script (m4-ccas).

Part 3 is not implemented yet in this patch, as it is a WIP.

Additionally, the mpn-generated moz.build files will need to be
generated for each platform, and stored in platform-specific
directories. (Eg: mpn/moz.build should really be mpn/linux.mozbuild, and
mpn/moz.build can include the right file based on the platform). I was
planning to get Linux working first and then extending to other
platforms.

MozReview-Commit-ID: ARnSpbeoaDM